feat: add support for multi-value attributes
Fully backwards compatible feature enhancement.

The RDN sequence for both subject and issuer is allowed to contain the
same attribute more than once. For example, Entrust certificates have
two OU attributes. Passing the new `multivalue: true` option to either
of the parse functions yields this output, where values for the subject
and issuer are now lists. In the example below, you can see both of the
OU values. Further, the aggregated string also includes both now.

    %{
      aggregated: "/C=US/CN=Entrust Certification Authority - L1K/O=Entrust, Inc./OU=See www.entrust.net/legal-terms, (c) 2012 Entrust, Inc. - for authorized use only",
      CN: ["Entrust Certification Authority - L1K"],
      C: ["US"],
      L: [],
      ST: [],
      O: ["Entrust, Inc."],
      OU: ["See www.entrust.net/legal-terms",
       "(c) 2012 Entrust, Inc. - for authorized use only"],
      emailAddress: []
    }

Without the `multivalue` option, it parses is it did prior to this
change dropping all but the last of the duplicated attributes, which is
broken technically. In the spirit of backwards compatibility, however,
it parses the same as before:

    %{
      aggregated: "/C=US/CN=Entrust Certification Authority - L1K/O=Entrust, Inc./OU=(c) 2012 Entrust, Inc. - for authorized use only",
      emailAddress: nil,
      CN: "Entrust Certification Authority - L1K",
      OU: "(c) 2012 Entrust, Inc. - for authorized use only",
      O: "Entrust, Inc.",
      ST: nil,
      L: nil,
      C: "US"
    }